+++ src/jsdoc-defaults.js
@@ -4,5 +4,5 @@ const path = require('path');
 
 module.exports = {
   configFile: path.join(__dirname, '../.jsdocrc'),
-  executablePath: path.join(require.resolve('jsdoc/jsdoc'), '../../.bin/')
+  packagePath: path.join(require.resolve('jsdoc/jsdoc'), '../')
 };